Remove FilterBoxListView and merge filter modules
Nothing uses FilterBoxListView since it is better to filter
within the data clerk.
